requirements.txt
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=1.1.0
langgraph-checkpoint>=4.0.0
tavily-python>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
